Manejamos css y js con webpack
