+++ public/js/main.js
@@ -1,92 +1,161 @@
-$(function() {
+/*
+ * ATTENTION: The "eval" devtool has been used (maybe by default in mode: "development").
+ * This devtool is neither made for production nor for readable output files.
+ * It uses "eval()" calls to create a separate source file in the browser devtools.
+ * If you are trying to read the output file, select a different devtool (https://webpack.js.org/configuration/devtool/)
+ * or disable the default devtool with "devtool: false".
+ * If you are looking for production-ready output files, see mode: "production" (https://webpack.js.org/configuration/mode/).
+ */
+/******/ (() => { // webpackBootstrap
+/******/ 	var __webpack_modules__ = ({
 
-    //Apertura y cierre del Menu lateral
-    $('.open-menu, .dismiss, .open-menu-int').click(function(){
-        $('.sidebar').toggleClass('fliph');
-        $('.open-menu-int').toggleClass('d-none');
-        $('.fa-caret-right').toggleClass('d-none');
-    });
+/***/ "./src/index.js":
+/*!**********************!*\
+  !*** ./src/index.js ***!
+  \**********************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-    /**-------------------------------------------------
-     Codigo para poner en modal los videos
-     -------------------------------------------*/
-    // Gets the video src from the data-src on each button
-    var $videoSrc;
-    $('.video-btn').click(function() {
-        $videoSrc = $(this).data( "src" );
-    });
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _css_menu_css__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./css/menu.css */ \"./src/css/menu.css\");\n/* harmony import */ var _css_bloquescontenido_css__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./css/bloquescontenido.css */ \"./src/css/bloquescontenido.css\");\n/* harmony import */ var _css_main_css__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./css/main.css */ \"./src/css/main.css\");\n/* harmony import */ var _js_main_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./js/main.js */ \"./src/js/main.js\");\n/* harmony import */ var _js_main_js__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(_js_main_js__WEBPACK_IMPORTED_MODULE_3__);\n/* harmony import */ var _js_plugins_js__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./js/plugins.js */ \"./src/js/plugins.js\");\n/* harmony import */ var _js_plugins_js__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(_js_plugins_js__WEBPACK_IMPORTED_MODULE_4__);\n/* harmony import */ var _js_checkRadioHoney_jquery__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./js/checkRadioHoney.jquery */ \"./src/js/checkRadioHoney.jquery.js\");\n/* harmony import */ var _js_checkRadioHoney_jquery__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(_js_checkRadioHoney_jquery__WEBPACK_IMPORTED_MODULE_5__);\n\n\n\n\n\n\n\n//# sourceURL=webpack://tutorial-estrategias/./src/index.js?");
 
-    // when the modal is opened autoplay it
-    $('#VideosModal').on('shown.bs.modal', function (e) {
-        // set the video src to autoplay and not to show related video. Youtube related video is like a box of chocolates... you never know what you're gonna get
-        $("#video").attr('src',$videoSrc + "?autoplay=1&amp;modestbranding=1&amp;showinfo=0" );
-    });
+/***/ }),
 
-    // stop playing the youtube video when I close the modal
-    $('#VideosModal').on('hide.bs.modal', function (e) {
-        // a poor man's stop video
-        $("#video").attr('src',$videoSrc);
-    });
+/***/ "./src/js/checkRadioHoney.jquery.js":
+/*!******************************************!*\
+  !*** ./src/js/checkRadioHoney.jquery.js ***!
+  \******************************************/
+/***/ (() => {
 
-    $('.flash-modal').modal('show');
-});
+eval("// the semi-colon before function invocation is a safety net against concatenated\n// scripts and/or other plugins which may not be closed properly.\n;\n\n(function ($, window, document, undefined) {\n  \"use strict\"; // undefined is used here as the undefined global variable in ECMAScript 3 is\n  // mutable (ie. it can be changed by someone else). undefined isn't really being\n  // passed in so we can ensure the value of it is truly undefined. In ES5, undefined\n  // can no longer be modified.\n  // window and document are passed through as local variables rather than global\n  // as this (slightly) quickens the resolution process and can be more efficiently\n  // minified (especially when both are regularly referenced in your plugin).\n  // Create the defaults once\n\n  var pluginName = \"checkRadioHoney\",\n      defaults = {\n    propertyName: \"value\"\n  }; // The actual plugin constructor\n\n  function Plugin(element, options) {\n    this.element = element; // jQuery has an extend method which merges the contents of two or\n    // more objects, storing the result in the first object. The first object\n    // is generally empty as we don't want to alter the default options for\n    // future instances of the plugin\n\n    this.settings = $.extend({}, defaults, options);\n    this._defaults = defaults;\n    this._name = pluginName;\n    this.init();\n  } // Avoid Plugin.prototype conflicts\n\n\n  $.extend(Plugin.prototype, {\n    init: function () {\n      // Place initialization logic here\n      // You already have access to the DOM element and\n      // the options via the instance, e.g. this.element\n      // and this.settings\n      // you can add more functions like the one below and\n      // call them like the example below\n      //this.yourOtherFunction( \"jQuery Boilerplate\" );\n      var elementos = $(\".pagina\", this.element);\n      var ultimo = elementos.hide().first().fadeIn(300);\n      var terminar = $(\"#terminar\", this.element).hide();\n      var progress = $(\".progress-bar\");\n      var avanceCuest = 12.5;\n      var contador = 1;\n      progress.css('width', avanceCuest + '%');\n      $(\"#avanzar\", this.element).on(\"click\", function () {\n        var nradios = $(\"[type=radio]\", ultimo).length / 2;\n        var nradiosCheck = $(\"[type=radio]:checked\", ultimo).length;\n\n        if (nradios == nradiosCheck) {\n          ultimo.hide();\n          ultimo = ultimo.next().fadeIn(300);\n          contador++;\n          avanceCuest = avanceCuest + 12.5;\n          progress.css('width', avanceCuest + '%');\n\n          if (contador == 8) {\n            $(this).hide();\n            terminar.show();\n          }\n        } else {\n          $('#HoneyError').modal('show');\n        }\n      });\n    },\n    getRadioButton: function (element) {\n      var name = $(element).attr(\"name\");\n      return $(\"[type=radio]\", element).length;\n    },\n    getRadioButtonChecked: function (element) {\n      var name = $(element).attr(\"name\");\n      return $(\"[type=radio]:checked\", element).length;\n    },\n    yourOtherFunction: function (text) {\n      // some logic\n      $(this.element).text(text);\n    }\n  }); // A really lightweight plugin wrapper around the constructor,\n  // preventing against multiple instantiations\n\n  $.fn[pluginName] = function (options) {\n    return this.each(function () {\n      if (!$.data(this, \"plugin_\" + pluginName)) {\n        $.data(this, \"plugin_\" + pluginName, new Plugin(this, options));\n      }\n    });\n  };\n})(jQuery, window, document);\n\n//# sourceURL=webpack://tutorial-estrategias/./src/js/checkRadioHoney.jquery.js?");
 
+/***/ }),
 
-// Shorthand for $( document ).ready()
-//Cuestionario Honey-Alonso, paginacion
-/*$(function() {
-    var cuestionario = $("#honeyAlonso");
-    var progress = $(".progress-bar");
-    var elementos = $(".pagina", cuestionario);
-    var ultimo = elementos.hide().first().fadeIn(300);
-    var terminar = $("#terminar", cuestionario).hide();
-    var contador = 1;
-    var avanceCuest = 12.5;
+/***/ "./src/js/main.js":
+/*!************************!*\
+  !*** ./src/js/main.js ***!
+  \************************/
+/***/ (() => {
 
-    progress.css('width', avanceCuest + '%');
+eval("$(function () {\n  //Apertura y cierre del Menu lateral\n  $('.open-menu, .dismiss, .open-menu-int').click(function () {\n    $('.sidebar').toggleClass('fliph');\n    $('.open-menu-int').toggleClass('d-none');\n    $('.fa-caret-right').toggleClass('d-none');\n  });\n  /**-------------------------------------------------\n   Codigo para poner en modal los videos\n   -------------------------------------------*/\n  // Gets the video src from the data-src on each button\n\n  var $videoSrc;\n  $('.video-btn').click(function () {\n    $videoSrc = $(this).data(\"src\");\n  }); // when the modal is opened autoplay it\n\n  $('#VideosModal').on('shown.bs.modal', function (e) {\n    // set the video src to autoplay and not to show related video. Youtube related video is like a box of chocolates... you never know what you're gonna get\n    $(\"#video\").attr('src', $videoSrc + \"?autoplay=1&amp;modestbranding=1&amp;showinfo=0\");\n  }); // stop playing the youtube video when I close the modal\n\n  $('#VideosModal').on('hide.bs.modal', function (e) {\n    // a poor man's stop video\n    $(\"#video\").attr('src', $videoSrc);\n  });\n  $('.flash-modal').modal('show');\n}); // Shorthand for $( document ).ready()\n//Cuestionario Honey-Alonso, paginacion\n\n/*$(function() {\n    var cuestionario = $(\"#honeyAlonso\");\n    var progress = $(\".progress-bar\");\n    var elementos = $(\".pagina\", cuestionario);\n    var ultimo = elementos.hide().first().fadeIn(300);\n    var terminar = $(\"#terminar\", cuestionario).hide();\n    var contador = 1;\n    var avanceCuest = 12.5;\n\n    progress.css('width', avanceCuest + '%');\n\n    $(\"#honeyAlonso #avanzar\").on( \"click\", function() {\n        ultimo.hide();\n        ultimo = ultimo.next().fadeIn(300);\n        contador++;\n        avanceCuest = avanceCuest + 12.5;\n        progress.css('width',avanceCuest + '%');\n\n        if(contador == 8){\n            $(this).hide();\n            terminar.show();\n        }\n    });\n});*/\n\n/*$(function() {\n    $(\"form\").each(function(){\n        $form = this;\n        $(\":submit\", this).click(function(e){\n\n            var names = {};\n\n            $(':radio', $form).each(function() {\n                names[$(this).attr('name')] = true;\n            });\n            var count = 0;\n            $.each(names, function() {\n                console.log(names);\n                count++;\n            });\n            console.log(count);\n            if (count > 0) {\n                e.preventDefault();\n                if ($(':radio:checked', $form).length < count) {\n                    console.log($(':radio:checked', $form).length);\n                    alert(\"falta contestar alguno\");\n                }else{\n                    $($form).submit();\n                }\n            }else{\n                $($form).submit();\n            }\n\n        });\n    });\n});*/\n\n//# sourceURL=webpack://tutorial-estrategias/./src/js/main.js?");
 
-    $("#honeyAlonso #avanzar").on( "click", function() {
-        ultimo.hide();
-        ultimo = ultimo.next().fadeIn(300);
-        contador++;
-        avanceCuest = avanceCuest + 12.5;
-        progress.css('width',avanceCuest + '%');
+/***/ }),
 
-        if(contador == 8){
-            $(this).hide();
-            terminar.show();
-        }
-    });
-});*/
+/***/ "./src/js/plugins.js":
+/*!***************************!*\
+  !*** ./src/js/plugins.js ***!
+  \***************************/
+/***/ (() => {
 
-/*$(function() {
-    $("form").each(function(){
-        $form = this;
-        $(":submit", this).click(function(e){
+eval("// Avoid `console` errors in browsers that lack a console.\n(function () {\n  var method;\n\n  var noop = function () {};\n\n  var methods = ['assert', 'clear', 'count', 'debug', 'dir', 'dirxml', 'error', 'exception', 'group', 'groupCollapsed', 'groupEnd', 'info', 'log', 'markTimeline', 'profile', 'profileEnd', 'table', 'time', 'timeEnd', 'timeline', 'timelineEnd', 'timeStamp', 'trace', 'warn'];\n  var length = methods.length;\n  var console = window.console = window.console || {};\n\n  while (length--) {\n    method = methods[length]; // Only stub undefined methods.\n\n    if (!console[method]) {\n      console[method] = noop;\n    }\n  }\n})(); // Place any jQuery/helper plugins in here.\n\n\n$(window).on(\"load\", function (e) {\n  $('.preloader').fadeOut('slow');\n});\n$(function () {\n  //Woo\n  new WOW().init();\n}); // Shorthand for $( document ).ready()\n\n//# sourceURL=webpack://tutorial-estrategias/./src/js/plugins.js?");
 
-            var names = {};
+/***/ }),
 
-            $(':radio', $form).each(function() {
-                names[$(this).attr('name')] = true;
-            });
-            var count = 0;
-            $.each(names, function() {
-                console.log(names);
-                count++;
-            });
-            console.log(count);
-            if (count > 0) {
-                e.preventDefault();
-                if ($(':radio:checked', $form).length < count) {
-                    console.log($(':radio:checked', $form).length);
-                    alert("falta contestar alguno");
-                }else{
-                    $($form).submit();
-                }
-            }else{
-                $($form).submit();
-            }
+/***/ "./src/css/bloquescontenido.css":
+/*!**************************************!*\
+  !*** ./src/css/bloquescontenido.css ***!
+  \**************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
-        });
-    });
-});*/
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n// extracted by mini-css-extract-plugin\n\n\n//# sourceURL=webpack://tutorial-estrategias/./src/css/bloquescontenido.css?");
+
+/***/ }),
+
+/***/ "./src/css/main.css":
+/*!**************************!*\
+  !*** ./src/css/main.css ***!
+  \**************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n// extracted by mini-css-extract-plugin\n\n\n//# sourceURL=webpack://tutorial-estrategias/./src/css/main.css?");
+
+/***/ }),
+
+/***/ "./src/css/menu.css":
+/*!**************************!*\
+  !*** ./src/css/menu.css ***!
+  \**************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n// extracted by mini-css-extract-plugin\n\n\n//# sourceURL=webpack://tutorial-estrategias/./src/css/menu.css?");
+
+/***/ })
+
+/******/ 	});
+/************************************************************************/
+/******/ 	// The module cache
+/******/ 	var __webpack_module_cache__ = {};
+/******/ 	
+/******/ 	// The require function
+/******/ 	function __webpack_require__(moduleId) {
+/******/ 		// Check if module is in cache
+/******/ 		var cachedModule = __webpack_module_cache__[moduleId];
+/******/ 		if (cachedModule !== undefined) {
+/******/ 			return cachedModule.exports;
+/******/ 		}
+/******/ 		// Create a new module (and put it into the cache)
+/******/ 		var module = __webpack_module_cache__[moduleId] = {
+/******/ 			// no module.id needed
+/******/ 			// no module.loaded needed
+/******/ 			exports: {}
+/******/ 		};
+/******/ 	
+/******/ 		// Execute the module function
+/******/ 		__webpack_modules__[moduleId](module, module.exports, __webpack_require__);
+/******/ 	
+/******/ 		// Return the exports of the module
+/******/ 		return module.exports;
+/******/ 	}
+/******/ 	
+/************************************************************************/
+/******/ 	/* webpack/runtime/compat get default export */
+/******/ 	(() => {
+/******/ 		// getDefaultExport function for compatibility with non-harmony modules
+/******/ 		__webpack_require__.n = (module) => {
+/******/ 			var getter = module && module.__esModule ?
+/******/ 				() => (module['default']) :
+/******/ 				() => (module);
+/******/ 			__webpack_require__.d(getter, { a: getter });
+/******/ 			return getter;
+/******/ 		};
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/define property getters */
+/******/ 	(() => {
+/******/ 		// define getter functions for harmony exports
+/******/ 		__webpack_require__.d = (exports, definition) => {
+/******/ 			for(var key in definition) {
+/******/ 				if(__webpack_require__.o(definition, key) && !__webpack_require__.o(exports, key)) {
+/******/ 					Object.defineProperty(exports, key, { enumerable: true, get: definition[key] });
+/******/ 				}
+/******/ 			}
+/******/ 		};
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/hasOwnProperty shorthand */
+/******/ 	(() => {
+/******/ 		__webpack_require__.o = (obj, prop) => (Object.prototype.hasOwnProperty.call(obj, prop))
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/make namespace object */
+/******/ 	(() => {
+/******/ 		// define __esModule on exports
+/******/ 		__webpack_require__.r = (exports) => {
+/******/ 			if(typeof Symbol !== 'undefined' && Symbol.toStringTag) {
+/******/ 				Object.defineProperty(exports, Symbol.toStringTag, { value: 'Module' });
+/******/ 			}
+/******/ 			Object.defineProperty(exports, '__esModule', { value: true });
+/******/ 		};
+/******/ 	})();
+/******/ 	
+/************************************************************************/
+/******/ 	
+/******/ 	// startup
+/******/ 	// Load entry module and return exports
+/******/ 	// This entry module can't be inlined because the eval devtool is used.
+/******/ 	var __webpack_exports__ = __webpack_require__("./src/index.js");
+/******/ 	
+/******/ })()
+;
