+++ src/css/main.css
@@ -924,7 +924,8 @@ position: relative;
 list-style: none;
 }
 
-.paso-paso ol li:nth-child(odd):before,.paso-paso ol li:nth-child(even):before {
+.paso-paso ol li:nth-child(odd):before,
+.paso-paso ol li:nth-child(even):before {
 content: counter(li);
 display: flex;
 -ms-flex-pack: center!important;
@@ -986,16 +987,14 @@ border-bottom-left-radius:0;
 /* rounded  list style                    */
 
 /* list */
-.rounded-list ol
-{
+.rounded-list ol{
 counter-reset:li; /* Initiate a counter */
 margin-left:0; /* Remove the default left margin */
 padding-left:0; /* Remove the default left padding */
 }
 
 /* item  */
-.rounded-list ol > li
-{
+.rounded-list ol > li {
 position:relative; /* Create a positioning context */
 list-style:none; /* Disable the normal item numbering */
 margin:0; /* Give each list item a left margin to make room for the numbers */
@@ -1005,8 +1004,7 @@ padding-top:0px;
 }
 
 /* number  */
-.rounded-list ol > li p:before
-{
+.rounded-list ol > li p:before{
 content: counter(li);
 counter-increment: li;
 position: absolute;
@@ -1119,7 +1117,7 @@ color: #fff;
 	float: right;
 }
 
-/*-----Bibliografía-----*/
-#bibliografia .list-group-flush > .list-group-item{
+/* -----Bibliografia----- */
+#bibliografia .list-group-flush > .list-group-item {
     overflow: hidden;
 }
