+++ src/index.js
@@ -1,0 +1,8 @@
+import "./css/menu.css";
+import "./css/bloquescontenido.css";
+import "./css/main.css";
+
+import "./js/main.js";
+import "./js/plugins.js";
+import "./js/checkRadioHoney.jquery";
+
